remove: memo_utils.jsを削除、中身の関数を一度memo.jsに移動
