+++ 04.class/memo.js
@@ -5,8 +5,8 @@ import { stdin, stdout } from "node:process";
 
 import minimist from "minimist";
 import sqlite3 from "sqlite3";
+import inquirer from "inquirer";
 
-import { closeDatabase, getMemoRows, chooseMemo } from "./memo_utils.js";
 import dbPromise from "./db_utils.js";
 
 class MemoApp {
@@ -100,6 +100,43 @@ class MemoApp {
     }
     await closeDatabase(this.db);
   }
+}
+
+async function closeDatabase(db) {
+  try {
+    await promise.close(db);
+  } catch (err) {
+    if (err instanceof Error && err.code === "SQLITE_MISUSE") {
+      console.error(err.message);
+    } else {
+      throw err;
+    }
+  }
+}
+
+async function getMemoRows(db) {
+  let memoRows;
+  try {
+    memoRows = await promise.all(db, "SELECT id, text FROM memo");
+  } catch (err) {
+    if (err instanceof Error && err.code === "SQLITE_ERROR") {
+      console.error(err.message);
+    } else {
+      throw err;
+    }
+  }
+  return memoRows;
+}
+
+async function chooseMemo(memoRows, name, type, message) {
+  const choices = memoRows.map((row) => ({
+    name: row.text.split("\n")[0],
+    value: row,
+  }));
+
+  const answers = await inquirer.prompt([{ name, type, message, choices }]);
+
+  return answers;
 }
 
 const promise = new dbPromise();
